test(integrations): split redelivery regression tests into their own file
Keeps the lifecycle test file under the repo-wide 600-line cap.
